+++ NEWS.md
@@ -15,6 +15,7 @@
 * Importing Records from BNE z39.50 - record not available error [MODINV-1395](https://folio-org.atlassian.net/browse/MODINV-1395)
 * Fix match results not passed to subsequent matches for consortium tenants [MODINV-1400](https://folio-org.atlassian.net/browse/MODINV-1400)
 * Fix missing logging after dependency bump [MODINV-1411](https://folio-org.atlassian.net/browse/MODINV-1411)
+* Improve Kafka header handling with case-insensitive TreeMap [MODINV-1414](https://folio-org.atlassian.net/browse/MODINV-1414)
 
 ### Tech Dept
 * Description ([ISSUE](https://folio-org.atlassian.net/browse/ISSUE))